Add files via upload
The Healthy Programmer Project is a Python program that aims to promote healthy habits for programmers who spend long hours in front of their computers. The program reminds the user to take breaks and engage in healthy activities such as drinking water, doing eye exercises, and physical activity. It runs continuously in the background and prompts the user at specified intervals. The program also logs each activity in a text file for future reference. The Healthy Programmer Project is a useful tool for programmers who want to stay healthy while working long hours on their computers. You can modify the audio and the text in this program according to your wish. This is easy to understand.